Ensure no partial results are used va the DnsCache.
Motivation:

We need to ensure we only use the DnsCache from the EventLoop when try to resolve a query with it. This is because otherwise we may see only partial results for a previous query as we are not done adding everything. Beside this we also should delay commiting any cached entries until we have completed the query.

Modifications:

- Delay commiting any entry to the cache until the query is complete
- Ensure we acces the cache only from the EventLoop to never see partial results

Result:

Fixes https://github.com/netty/netty/issues/7882